Undo a deletion properly, and hand the debt back with it
- Restore now drops the sync tombstone the delete trigger left behind and
  stamps a fresh updatedAt, so an undone record is not deleted again on the
  next sync; the undo snackbar also stays up long enough to be pressed.
- A swiped row that comes back is no longer stuck in its dismissed state:
  the swipe state is remembered per composition instead of saved per key.
- Transactions remember which debt they moved (debtId + delta, migration 19).
  Deleting a repayment gives the amount back and reopens the debt, deleting
  the record that opened a debt takes the debt with it, and undo puts both
  sides back exactly as they were, creation date included.
- Deletion goes through one repository call that gathers items, reverses the
  debt and reports the record, instead of every screen doing its own half.
- Instrumented tests cover the three debt shapes and the tombstone.

+++ app/src/main/java/by/mlastovsky/kosht/ui/components/UndoDeleteSnackbar.kt
@@ -31,7 +31,7 @@ fun UndoDeleteSnackbar(
             val result = hostState.showSnackbar(
                 message = message,
                 actionLabel = undo,
-                duration = SnackbarDuration.Short
+                duration = SnackbarDuration.Long
             )
             if (result == SnackbarResult.ActionPerformed) {
                 viewModel.restore(deleted)
